Split spaceship code into packages and files
The atom has been split, I repeat, the atom has been split!
